server/requirements.txt
google
apache-libcloud==2.8.1
arpy==1.1.1
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
config==0.5.0
defusedxml==0.6.0
docopt==0.6.2
filedepot==0.7.1
firebase-admin==4.0.0
Flask==1.1.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lockfile==0.12.2
lxml==4.5.0
MarkupSafe==1.1.1
msgpack==1.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-gnupg==0.4.5
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
